backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic>=2.0.0
python-dotenv==1.0.0

# Database
mysql-connector-python==8.3.0

# Audio Processing & Transcription
openai-whisper==20231117
pydub==0.25.1
soundfile==0.12.1
librosa==0.10.1

# AI Services
google-genai

# API Requests
requests==2.31.0